CUMULUS-2309 move test data creation to api package
