+++ packages/api/tests/lib/test-granule-delete.js
@@ -39,7 +39,7 @@ const { deleteGranuleAndFiles } = require('../../lib/granule-delete');
 
 const { migrationDir } = require('../../../../lambdas/db-migration');
 
-const { createGranuleAndFiles } = require('../db-data-helpers/create-test-data');
+const { createGranuleAndFiles } = require('../../lib/create-test-data');
 
 const testDbName = `granules_${cryptoRandomString({ length: 10 })}`;
